// packages/client/src/pages/UserEntry.model.ts
export enum AgeRangeType {
  from18To24 = "18-24",
  from25To34 = "25-34",
  from35To44 = "35-44",
  from45To54 = "45-54",
  from55To64 = "55-64",
  from65To74 = "65-74",
}

export enum SexType {
  MALE = "m",
  FEMALE = "f",
}

export enum TimeDayType {
  DAY = "day",
  EVENING = "evening",
  MORNING = "morning",
  NIGHT = "night",
}